+++ frontend/src/pages/Profile/ProfilePage.jsx
@@ -51,7 +51,7 @@ export const ProfilePage = () => {
 
   return (
     <div className="bg-slate-100 min-h-screen">
-      <ProfileHeader profileInfo={profileInfo} profileOwner={profileOwner} />
+      <ProfileHeader profileInfo={profileInfo} profileOwner={profileOwner} token={token} setToken={setToken}/>
       <div className="md:flex gap-4 md:max-w-screen-lg w-screen mx-auto px-4">
         <div className="flex-auto basis-5/12">
           <Intro profileInfo={profileInfo} profileOwner={profileOwner} token={token} setToken={setToken}/>
